+++ package.json
@@ -21,7 +21,8 @@
 	},
 	"files": [
 		"index.js",
-		"index.d.ts"
+		"index.d.ts",
+		"lib"
 	],
 	"keywords": [
 		"get",
